tests/interop/requirements.txt
eclipse-zenoh==1.0.0
